+++ src/wrapper-mobile.js
@@ -1,2 +1,0 @@
-define('converse-minimize', [], function () { return; });
-define('converse-dragresize', [], function () { return; });